% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{generate_trajs}
\alias{generate_trajs}
\title{Use value iteration to solve a 5x5 gridworld}
\usage{
generate_trajs(policy, obstacles, wind)
}
\description{
Use value iteration to solve a 5x5 gridworld
}
