% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{value_iteration}
\alias{value_iteration}
\title{Use value iteration to solve a 5x5 gridworld}
\usage{
value_iteration(reward, obstacles, wind, beta)
}
\description{
Use value iteration to solve a 5x5 gridworld
}
